Se agrego dependencia a winstone para loggear errores, se agrego la capa services y se desacoplo codigo del archivo usersController.js
