+++ src/config/winstone.js
@@ -1,0 +1,14 @@
+const winston = require('winston');
+
+
+const Logger = winston.createLogger({
+    transports:[
+        new winston.transports.File({
+            filename: 'app.log',
+            level: 'error'
+        })
+    ]
+});
+
+
+exports.module = Logger;
